Endrer til å bruke hentUtvidetBarnetrygd repometode
