+++ src/main/kotlin/no/nav/familie/ba/infotrygd/repository/UtbetalingRepository.kt
@@ -2,7 +2,7 @@ package no.nav.familie.ba.infotrygd.repository
 
 import no.nav.commons.foedselsnummer.FoedselsNr
 import no.nav.familie.ba.infotrygd.model.db2.Utbetaling
-import no.nav.familie.ba.infotrygd.model.dl1.Stønad
+import no.nav.familie.ba.infotrygd.model.dl1.TrunkertStønad
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.stereotype.Repository
@@ -30,14 +30,5 @@ interface UtbetalingRepository : JpaRepository<Utbetaling, Long> {
         AND utbet.virksomFom = :#{#stonad.virkningFom}
         AND utbet.utbetalingstype = 'M'
     """)
-    fun hentUtbetalingerByStønad(stonad: Stønad): List<Utbetaling>
-
-    @Query("""
-        SELECT utbet FROM Utbetaling utbet
-        WHERE utbet.personKey = :#{#stonad.personKey}   
-        AND utbet.startUtbetalingMåned = :#{#stonad.iverksattFom}
-        AND utbet.virksomFom = :#{#stonad.virkningFom}
-        AND utbet.utbetalingstype = 'M'
-    """)
-    fun hentUtbetalingerByTrunkertStønad(stonad: TrunkertStønad): List<Utbetaling>
+    fun hentUtbetalingerByStønad(stonad: TrunkertStønad): List<Utbetaling>
 }
